requirements.txt
Flask==3.0.3
google-genai==0.5.0
python-dotenv==1.0.1
gunicorn==22.0.0
pandas==2.2.2